Full Object-Oriented Classes
Now there are 2 classes, one of which holds objects, and then the main class which will do display functions...
